Changed const.calcFFT(), modified nonlinfit.get(), added progress bar
const.calcFFT() was calculating a one-sided FFT.  I have changed it
so that it now calculates a two-sided FFT (although it only displays
the one side).  This amounts to removing a factor of 2 for all freq-
uencies not equal to 0.

nonlinfit.get() now has a second argument that specifies whether to
get the parameter value or parameter error or both.

Added the textprogressbar.m function which generates a text-based
progress bar that uses only one line.  This needs some work as it
doesn't handle crashes very well due to its use of a persistent
variable