Multiple changes to functions. Major calculation change to FFT stuff
Changed the scaling of the const.calcFFT function so that it reports
the correct one-sided FFT.  A warning is now given if complex data
is provided, as the FFT can no longer be considered only one-sided.
Now, trapz(f,PSD) = var(x) where x is the time-series data and
PSD is the calculated power spectral density from const.calcPSD(x,dt).

Some additional changes to fitting functions, mostly for plotting,
as well as a non-negative fit for linear fitting.

A couple of plotting functions were added